more reorganization
still deciding what's best
